0.4.29: saturate exp downscale + ceiling guards (Maniference O26 fix)
Root cause: downscale_q64_to_q32 wrapped oversized Q64.64 results
(including the exp overflow sentinel i128::MAX) via a plain `as i64`
cast. On realtime/compact profiles, exp past the wrap threshold came
back as arbitrary garbage. Measured blast radius at Q22.10 (verified
pre-fix in a worktree): fused::silu(-70) returned -70 — the input
passed through unsquashed, Maniference's Mixtral 200x residual spike —
and tanh(25) returned -1 (sign flip). Latent 15 months because dense
LLM gate values stay within +/-30; Mixtral expert gates reach +/-70.

Fixes, all at the defect seam or the adds it feeds:
- downscale_q64_to_q32 saturates to i64::MAX/MIN; oversized results
  stay at the ceiling so storage downscales detect them.
- Ceiling guards where a follow-up add on a saturated exp would wrap
  (or panic in debug): fused::silu -> 0, FixedPoint::tanh -> 1 (both
  correctly rounded for every saturating input), FixedPoint::cosh ->
  loud "cosh overflow", FASC evaluate_tanh -> 1, FASC evaluate_cosh ->
  TierOverflow, sinhcosh_at_compute_tier pins the cosh sum at the
  ceiling. New shared helper compute_ceiling() names the value.
- Audited remaining exp_at_compute_tier consumers: softmax and
  softmax_mix subtract the max first (exp inputs <= 0) — safe; sinh
  overflows storage before exp saturates on every profile — safe.
- tq19_bench weight generator: i64 arithmetic (i32 overflowed at
  4096x4096 in debug — pre-existing, test infra only).

Regression tests: exp monotonicity property (the wrap broke it), silu
deep-negative -30..-100 all profiles, tanh saturation region = 1 all
profiles, try_tanh/try_cosh narrow-profile behavior.

Verified: Q22.10 fused 20/0 + linalg 53/0; Q16.16 20/0 + 53/0 + profile
suite; Q32.32 19/0 + 52/0 + compact suite; embedded full suite with
inference 1141/0; decimal contract harness (balanced, pinned corpus)
11/0; public-api-freshness PASS.

Reported-by: Maniference (NOTE_FROM_MANIFERENCE2.md, O26)

+++ tests/tq19_bench.rs
@@ -16,7 +16,8 @@ fn make_activations(n: usize) -> Vec<BinaryStorage> {
 }
 
 fn make_weights(rows: usize, cols: usize) -> Vec<i16> {
-    (0..rows * cols).map(|i| ((i as i32 * 137 - 5000) % (SCALE / 2)) as i16).collect()
+    // i64: at 4096×4096, i·137 exceeds i32::MAX (debug-mode overflow panic).
+    (0..rows * cols).map(|i| ((i as i64 * 137 - 5000) % (SCALE as i64 / 2)) as i16).collect()
 }
 
 fn make_trits(n: usize) -> Vec<i8> {
